Move flutter_archive.dart to tool/utils/
